[refactor] : dto 폴더 구조 조정
