+++ src/main/java/com/dia/dia_be/service/vip/impl/VipJournalServiceImpl.java
@@ -13,9 +13,9 @@ import com.dia.dia_be.domain.Product;
 import com.dia.dia_be.domain.QConsulting;
 import com.dia.dia_be.domain.QProduct;
 import com.dia.dia_be.domain.QScript;
-import com.dia.dia_be.dto.vip.ResponseJournalDTO;
-import com.dia.dia_be.dto.vip.ResponseJournalScriptDTO;
-import com.dia.dia_be.dto.vip.ResponseSimpleJournalDTO;
+import com.dia.dia_be.dto.vip.journalDTO.ResponseJournalDTO;
+import com.dia.dia_be.dto.vip.journalDTO.ResponseJournalScriptDTO;
+import com.dia.dia_be.dto.vip.journalDTO.ResponseSimpleJournalDTO;
 import com.dia.dia_be.exception.GlobalException;
 import com.dia.dia_be.exception.VipErrorCode;
 import com.dia.dia_be.repository.ConsultingRepository;
