+++ src/main/java/com/dia/dia_be/service/vip/intf/VipCategoryService.java
@@ -2,7 +2,7 @@ package com.dia.dia_be.service.vip.intf;
 
 import java.util.List;
 
-import com.dia.dia_be.dto.vip.ResponseCategoryGetDTO;
+import com.dia.dia_be.dto.vip.categoryDTO.ResponseCategoryGetDTO;
 
 public interface VipCategoryService {
