+++ src/main/java/com/dia/dia_be/service/vip/intf/VipJournalService.java
@@ -2,9 +2,9 @@ package com.dia.dia_be.service.vip.intf;
 
 import java.util.List;
 
-import com.dia.dia_be.dto.vip.ResponseJournalDTO;
-import com.dia.dia_be.dto.vip.ResponseJournalScriptDTO;
-import com.dia.dia_be.dto.vip.ResponseSimpleJournalDTO;
+import com.dia.dia_be.dto.vip.journalDTO.ResponseJournalDTO;
+import com.dia.dia_be.dto.vip.journalDTO.ResponseJournalScriptDTO;
+import com.dia.dia_be.dto.vip.journalDTO.ResponseSimpleJournalDTO;
 
 public interface VipJournalService {
