+++ src/main/java/com/dia/dia_be/service/vip/intf/VipNotificationService.java
@@ -2,12 +2,14 @@ package com.dia.dia_be.service.vip.intf;
 
 import java.util.List;
 
-import com.dia.dia_be.dto.vip.ResponseNotificationGetDTO;
+import com.dia.dia_be.dto.vip.notificationDTO.ResponseNotificationGetDTO;
 
 public interface VipNotificationService {
 
 	List<ResponseNotificationGetDTO> getNotifications(Long customerId);
+
 	void deleteAllNotifications(Long customerId);
+
 	void markAllNotificationsAsRead(Long customerId);
 
 }
